// src/app/shared/models/functional-checklist.ts
export class FunctionalChecklist {
    id: number;
    patientId: number;
    bathingShowering: FunctionalScore;
    bowelBladderManagement: FunctionalScore;
    dressing: FunctionalScore;
    eating: FunctionalScore;
    feeding: FunctionalScore;
    functionalMobility: FunctionalScore;
    personalDeviceCare: FunctionalScore;
    personalHygieneGrooming: FunctionalScore;
    sleepRest: FunctionalScore;
    sexualActivity: FunctionalScore;
    toiletHygiene: FunctionalScore;
    careOfOthers: FunctionalScore;
    careOfPets: FunctionalScore;
    parenting: FunctionalScore;
    communication: FunctionalScore;
    publicTransportUse: FunctionalScore;
    financialManagement: FunctionalScore;
    healthManagement: FunctionalScore;
    homeManagement: FunctionalScore;
    mealPreparationCleanup: FunctionalScore;
    medicationManagement: FunctionalScore;
    safetyProcedures: FunctionalScore;
    shopping: FunctionalScore;
    educationalExploration: FunctionalScore;
    formalEducationalParticipation: FunctionalScore;
    informalEducationalParticipation: FunctionalScore;
    employmentInterestsPursuits: FunctionalScore;
    seekingEmployment: FunctionalScore;
    jobPerformance: FunctionalScore;
    capacity: FunctionalScore;
    retirementAdjustment: FunctionalScore;
    seekingVolunteerEmployment: FunctionalScore;
    volunteer: FunctionalScore;
    participation: FunctionalScore;
    leisureExploration: FunctionalScore;
    leisureParticipation: FunctionalScore;
    family: FunctionalScore;
    peerFriend: FunctionalScore;
    community: FunctionalScore;
    religion: FunctionalScore;
    violenceAggression: FunctionalScore;
    suicide: FunctionalScore;
    selfNeglect: FunctionalScore;
    wandering: FunctionalScore;
    sexualOffences: FunctionalScore;
    inappropriateSexualBehaviour: FunctionalScore;
    deliberateFireSetting: FunctionalScore;
    deliberateSelfHarm: FunctionalScore;
    otherSelfHarm: FunctionalScore;
    abuseExploitationFromOthers: FunctionalScore;
    accidentalInjury: FunctionalScore;
    substanceMisuse: FunctionalScore;
    other: FunctionalScore;
    nonContactWithServices: FunctionalScore;
    nonEngagementWithCarePlan: FunctionalScore;
    nonComplianceWithMedication: FunctionalScore;
    
    additionalNotes: string;
    calculatedScore: number;
    totalScore: number;
    completedBy: number;
    completedDate: Date;
    nextReviewDate: Date;
    circulatedTo: string
    signOffBy: number;
    isSignOff: Boolean;
    constructor() {
        this.nextReviewDate = new Date(new Date().setMonth(new Date().getMonth() + 3));
        this.isSignOff = false;
        this.bathingShowering = 0;
        this.bowelBladderManagement = 0;
        this.dressing = 0;
        this.eating = 0;
        this.feeding = 0;
        this.functionalMobility = 0;
        this.personalDeviceCare = 0;
        this.personalHygieneGrooming = 0;
        this.sleepRest = 0;
        this.sexualActivity = 0;
        this.toiletHygiene = 0;
        this.careOfOthers = 0;
        this.careOfPets = 0;
        this.parenting = 0;
        this.communication = 0;
        this.publicTransportUse = 0;
        this.financialManagement = 0;
        this.healthManagement = 0;
        this.homeManagement = 0;
        this.mealPreparationCleanup = 0;
        this.medicationManagement = 0;
        this.safetyProcedures = 0;
        this.shopping = 0;
        this.educationalExploration = 0;
        this.formalEducationalParticipation = 0;
        this.informalEducationalParticipation = 0;
        this.employmentInterestsPursuits = 0;
        this.seekingEmployment = 0;
        this.jobPerformance = 0;
        this.capacity = 0;
        this.retirementAdjustment = 0;
        this.seekingVolunteerEmployment = 0;
        this.volunteer = 0;
        this.participation = 0;
        this.leisureExploration = 0;
        this.leisureParticipation = 0;
        this.family = 0;
        this.peerFriend = 0;
        this.community = 0;
        this.religion = 0;
        this.violenceAggression = 0;
        this.suicide = 0;
        this.selfNeglect = 0;
        this.wandering = 0;
        this.sexualOffences = 0;
        this.inappropriateSexualBehaviour = 0;
        this.deliberateFireSetting = 0;
        this.deliberateSelfHarm = 0;
        this.otherSelfHarm = 0;
        this.abuseExploitationFromOthers = 0;
        this.accidentalInjury = 0;
        this.substanceMisuse = 0;
        this.other = 0;
        this.nonContactWithServices = 0;
        this.nonEngagementWithCarePlan = 0;
        this.nonComplianceWithMedication = 0;

    }
}